+++ .gitignore
@@ -1,7 +1,28 @@
-node_modules/
+node_modules
 misc/coverage.html
 .vocrc
 v8.log
+tmp
+*.tgz
+*.[tT][xX][tT]
+*.[cC][sS][vV]
+*.[dD][iIbB][fF]
+*.[pP][rR][nN]
+*.[sS][lL][kK]
+*.socialcalc
+*.[xX][lL][sSwWcC]
+*.[xX][lL][sS][xXmMbB]
+*.[oO][dD][sS]
+*.[fF][oO][dD][sS]
+*.[xX][mM][lL]
+*.[uU][oO][sS]
+*.[wW][kKqQbB][S1234567890]
+*.[qQ][pP][wW]
+*.123
+*.htm
+*.html
+*.sheetjs
+*.exe
 perf.log
 ctest/test.js
 ctest/sauce*
