+++ .npmignore
@@ -1,8 +1,33 @@
 test/
+ctest/
+index.html
 misc/
-node_modules/
-misc/coverage.html
+node_modules
+*.tgz
+_book
+book.json
+tmp
+*.[tT][xX][tT]
+*.[cC][sS][vV]
+*.[dD][iIbB][fF]
+*.[pP][rR][nN]
+*.[sS][lL][kK]
+*.socialcalc
+*.[xX][lL][sSwWcC]
+*.[xX][lL][sS][xXmMbB]
+*.[oO][dD][sS]
+*.[fF][oO][dD][sS]
+*.[xX][mM][lL]
+*.[uU][oO][sS]
+*.[wW][kKqQbB][S1234567890]
+*.[qQ][pP][wW]
+*.123
+*.htm
+*.html
+*.sheetjs
+*.exe
 .gitignore
+.eslintrc
 .jshintrc
 Makefile
 .npmignore
@@ -14,4 +39,3 @@ bits/
 .vocrc
 v8.log
 perf.log
-ssf*.tgz
