+++ bits/01_version.js
@@ -1,1 +1,1 @@
-SSF.version = '0.9.0';
+SSF.version = '0.9.1';
