+++ bits/69_numfoot.js
@@ -1,0 +1,3 @@
+return function write_num(type/*:string*/, fmt/*:string*/, val/*:number*/)/*:string*/ {
+	return (val|0) === val ? write_num_int(type, fmt, val) : write_num_flt(type, fmt, val);
+};})();
