+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "ssf",
-	"version": "0.9.0",
+	"version": "0.9.1",
 	"author": "SheetJS",
 	"description": "Format data using ECMA-376 spreadsheet Format Codes",
 	"keywords": [ "format", "sprintf", "spreadsheet" ],
@@ -11,7 +11,7 @@
 	"dependencies": {
 		"voc":"",
 		"colors":"0.6.2",
-		"frac":"0.3.1"
+		"frac":"~1.0.6"
 	},
 	"devDependencies": {
 		"mocha":"",
