+++ test/oddities.json
@@ -31,7 +31,7 @@
   ["#,##0.000000000", [1000000, "1,000,000.000000000"]],
   ["#,###", [1, "1"], [-1, "-1"], [0,""], [12345.6789, "12,346"], ["TODO", "TODO"]],
   ["#.##", [1, "1."], [-1, "-1."], [0,"."], ["sheetjs", "sheetjs"]],
-  ["0;0", [1.1, "1"], [-1.1, "-1"], [0,"0"], ["sheetjs", "sheetjs"]],
+  ["0;0", [1.1, "1"], [-1.1, "1"], [0,"0"], ["sheetjs", "sheetjs"]],
   ["0.0", [1, "1.0"], [-1, "-1.0"], [0,"0.0"], ["sheetjs", "sheetjs"]],
   ["0.00", [1.0001, "1.00"], [-1, "-1.00"], [0,"0.00"], ["sheetjs", "sheetjs"]],
   ["0.000", [1, "1.000"], [-1, "-1.000"], [0,"0.000"], ["sheetjs", "sheetjs"]],
@@ -146,6 +146,14 @@
   ["☃", [0], [1], [-1]],
   ["#0#######", [12345, "012345"], [12345.4321, "012345"], [12345.6789, "012346"]],
   ["##,##", [12345, "12,345", ""], [12345.4321, "12,345", ""], [12345.6789, "12,346", ""]],
-  [0, [12345,"12345"], [4294967296.5, 4294967297]],
+  [0, [12345,"12345"], [4294967296.5, "4294967297"]],
+  ["\"Rs.\"#,##0.00", [-51968287, "-Rs.51,968,287.00"], [2000000, "Rs.2,000,000.00"]],
+  ["$#.00", [3.14159, "$3.14"], [-3.14159, "-$3.14"]],
+  ["\"This is a \".00\"test\"000", [-3.14159, "-This is a 3.14test159"], [3.14159, "This is a 3.14test159"]],
+  ["[$INR]\\ #,##0.00", [3.14159, "INR 3.14"], [-3.14159, "-INR 3.14"]],
+  ["[$₹-4009]\\ #,##0.00", [3.14159, "₹ 3.14"], [-3.14159, "-₹ 3.14"]],
+  ["[$£-809]#,##0.0000;\\-[$£-809]#,##0.0000", [3.14159, "£3.1416"], [-3.14159, "-£3.1416"]],
+  ["\"-\"0.00", [3.14159, "-3.14"], [-3.14159, "--3.14"]],
+  ["[$-409]mmm\\-yy", [12345, "Oct-33"]],
   ["\"foo\";\"bar\";\"baz\";\"qux\";\"foobar\"", [1], [0], [-1], ["sheetjs"]]
 ]
